小范围修正 model，增加自增和部分列的修正 测试和修正了一堆小 bug， 完善第一版，可以写入数据了 todo 待完善所有公式然后开搞

+++ shuang/src/test/test_for_formula.py
@@ -2,6 +2,7 @@
 __author__ = 'zr'
 import re
 import datetime
+
 import os
 import django
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "DJANGO1.settings")
@@ -10,7 +11,8 @@ django.setup()
 from shuang.src.basic.basic_model import FoliageSsq
 from shuang.src.basic.basic_model import TSsqShishibiao
 from shuang.src.formula.formula_model import ssq_formula
-
+from shuang.src.formula.formula_model import  ssq_formula_fact
+import time
 # ssq1=map(int,['05', '14', '20', '26', '30', '33', '12'])
 # print ssq1
 # print ssq1.reverse()
@@ -51,9 +53,9 @@ formula_type='get_blue_tail_list'
 
 
 ###因为我想做批量校验，减少数据库 io，所以希望输入的就是两个 ssq，但是在批量验证的时候，我可以通过批量输入公式，然后来进行校验。先循环 ssq，再循环 formula
-##如果输入的是ssq 范围和公式列表，则按一个一个 ssq 和公式列表来计算，todo num_range 要默认1，如何默认1？要校验 formula list 是否有效
+##如果输入的是ssq 范围和公式列表，则按一个一个 ssq 和公式列表来计算，
 
-###需要新增一个提交 batch 的计算界面。。。。差不多就是计算请求了，然后通过这个来计算请求的计算结果。
+###todo 需要新增一个提交 batch 的计算界面。。。。差不多就是计算请求了，然后通过这个来计算请求的计算结果。
 ###不过这个可以迟一点做，先做计算完才能停的，即计算完就显示计算结果这样。还有好多内容，还要显示计算结果。。。。不过这个似乎那个 django 的前段也可以做。够快速了。
 
 
@@ -61,9 +63,10 @@ formula_type='get_blue_tail_list'
 ###以下是校验数据的结果是否正确，并不需要写入数据库？还是写以下好吧？按 batch 记录，然后记录 batch 和提交的计算结果？
 ###todo 增加计算结果的 batch 字段，同时增加了写入数据库的内容  ；其实还是会有只算值，不保存结果的，比如测试一个公式
 def caculate_result(now,next_ssq,formula):
+    '''根据输入的第一个 ssq 和目标 ssq 以及公式 object 验证是否正确，并写入 batch'''
 ###计算结果值
     result_value=eval(formula.formula_express)
-    print 'result_value=',result_value,'now balls:',now.list_all_balls()
+    print 'result_value=',result_value,'now balls:',now.get_all_balls_byList()
     formula_value=[result_value]
     if formula.formula_type=='get_blue_tail_list': #取蓝球尾数的值
         if result_value >10:
@@ -79,34 +82,36 @@ def caculate_result(now,next_ssq,formula):
 ###运行公式计算公式的结果
 
     ###获得需要验证的下一个 ssq 或 通过快速的上去了验证结果
-    if formula_value :
+    if formula_value and next_ssq :
         if formula.formula_type=='get_blue_tail_list': #取蓝球尾数的值
             for i in formula_value:
                 if i==next_ssq.b1:
                     print 'i',i
                     print True,next_ssq.list_all_balls()
-                    return 'True'
+                    return 'True',formula_value
                     # todo insert into result,需要在循环结束以后
 
         elif formula.formula_type=='kill red':
             for i in formula_value:
                 if i==next_ssq.r1 or i==next_ssq.r2 or i==next_ssq.r3 or i==next_ssq.r4 or i==next_ssq.r5 or i==next_ssq.r6:
-                    print i,'Ture'
-                    return  'Ture'
-        return 'False'
+                    return  'False',formula_value #如果在下一期有就不是错了
+                else:
+                    return 'True', formula_value
+        return 'Null',formula_value ## 不属于任何一种类型，返回 null
     else :
-        return 'Null'
+        return 'Null',formula_value
 
 ###把公式结果和验证结果插入
 
 
-def formula_test(num_start,num_range,formula_list):#
-    '''输入起始期数，和范围（空则取指定ssq，0取所有值），以及公式id列表（空则计算所有公式）'''
-#获取 ssq 列表
+def formula_test(num_start,num_range,formula_list):
+    '''
+    num_start 起始期数， num_range 计算期数范围（空则取指定ssq，0取所有值），formula_list 公式id列表（空则计算所有公式）
+    '''
+
+#根据 range 获取 ssq 列表
     ssq_list = []
     if  num_range>1 :
-        #todo 根据 numstart 和 range 获得需要计算的，排好序的 ssq 序列，下面的取 ssq 要改，单纯的加一不靠谱。需要做好列表穿进去
-                    #todo 增加 formulalist 的取值，前端传进来的估计只有 formula 的 id 不一定能传值进来。
         ssq_list=TSsqShishibiao.objects.filter(num__gte=num_start)[0:num_range-1]  ##ssqlist，效率差点就差点吧
     elif num_range is None :
         ssq_list = TSsqShishibiao.objects.filter(num__exact=num_start)# 为空取指定的 ssq
@@ -124,12 +129,33 @@ def formula_test(num_start,num_range,formula_list):#
     else:
         formulas_list=ssq_formula.objects.filter(attribute2__exact='Active')
 
+    batch=int(time.time())
 #开始循环计算最终的校验结果
     print ssq_list
     print formulas_list
-    for p in ssq_list:
-        for j in  formulas_list:
-            result=caculate_result(p,p.get_next(),j)
+    for now_ssq in ssq_list:
+        for now_formula in  formulas_list:
+            next_ssq=now_ssq.get_next()
+            if next_ssq:
+                result,re_value=caculate_result(now_ssq,next_ssq,now_formula)#计算结果
+
+                fact=ssq_formula_fact()
+                fact.batch='%d' %batch
+                fact.now_periods=now_ssq.num
+                fact.target_periods=next_ssq.num
+                fact.result=result
+                fact.formula_value=','.join(map(str,re_value))#caculate_result的返回值里面还要增加返回计算的数据
+                fact.formula_type=now_formula.formula_type
+                fact.create_date=datetime.datetime.today()
+                fact.source_ssq=','.join(map(str,now_ssq.get_all_balls_byList()))
+                fact.target_ssq=','.join(map(str,next_ssq.get_all_balls_byList()))
+                fact.save()
             # print p.num,p.get_next().num,j.formula_express
+    return batch
 #
-formula_test(2003033,None,'')
+formula_test(2017016,10,'')
+
+
+##todo 完善所有公式，然后计算
+##todo 数据写进 controller 里面去
+##
